Timestamp u64 -> u32

+++ bee-message/src/output/feature_block/expiration_unix.rs
@@ -8,19 +8,19 @@ use bee_common::packable::{Packable, Read, Write};
 ///
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, derive_more::From)]
 #[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
-pub struct ExpirationUnixFeatureBlock(u64);
+pub struct ExpirationUnixFeatureBlock(u32);
 
 impl ExpirationUnixFeatureBlock {
     /// The feature block kind of an `ExpirationUnixFeatureBlock`.
     pub const KIND: u8 = 6;
 
     /// Creates a new `ExpirationUnixFeatureBlock`.
-    pub fn new(timestamp: u64) -> Self {
+    pub fn new(timestamp: u32) -> Self {
         timestamp.into()
     }
 
     /// Returns the timestamp.
-    pub fn timestamp(&self) -> u64 {
+    pub fn timestamp(&self) -> u32 {
         self.0
     }
 }
@@ -39,6 +39,6 @@ impl Packable for ExpirationUnixFeatureBlock {
     }
 
     fn unpack_inner<R: Read + ?Sized, const CHECK: bool>(reader: &mut R) -> Result<Self, Self::Error> {
-        Ok(Self::new(u64::unpack_inner::<R, CHECK>(reader)?))
+        Ok(Self::new(u32::unpack_inner::<R, CHECK>(reader)?))
     }
 }
